requirements.txt
bs4
requests
msal
asyncio
python-dotenv